+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/datastructure/resource/PersistentResource.java
@@ -17,39 +17,16 @@
  * under the License.
  */
 
-package org.apache.iotdb.commons.pipe.datastructure;
+package org.apache.iotdb.commons.pipe.datastructure.resource;
 
-public class Result<T, E> {
+import org.apache.iotdb.commons.consensus.index.ProgressIndex;
 
-  private final T ok;
-  private final E err;
+public interface PersistentResource {
+  ProgressIndex getProgressIndex();
 
-  private Result(final T ok, final E err) {
-    this.ok = ok;
-    this.err = err;
-  }
+  /** Used in PipeHistoricalDataRegionTsFileAndDeletionExtractor to sort PersistentResource */
+  long getFileStartTime();
 
-  public static <T, E> Result<T, E> ok(final T value) {
-    return new Result<>(value, null);
-  }
-
-  public static <T, E> Result<T, E> err(final E error) {
-    return new Result<>(null, error);
-  }
-
-  public boolean isOk() {
-    return ok != null;
-  }
-
-  public boolean isErr() {
-    return err != null;
-  }
-
-  public T getOk() {
-    return ok;
-  }
-
-  public E getErr() {
-    return err;
-  }
+  /** Used in PipeHistoricalDataRegionTsFileAndDeletionExtractor to sort PersistentResource */
+  long getFileEndTime();
 }
